CI: simplify notebook build
The used path filter dorny/paths-filter only offers glob patterns, while
we want to limit the conversion to the pattern in the Makefile. Instead
of re-implementing this with Some Other Github Action (tm), remove code
that checks for changed notebooks. Since we have only a handful of them,
just create them all. Also use the Makefile for that.